pyproject.toml
[tool.poetry]
name = "brotli-azf"
version = "1.0.0"
description = "Brotli encode/decode Azure Functions App"
authors = ["myl7 <myl@myl.moe>"]

[tool.poetry.dependencies]
python = "^3.9.1"
azure-functions = "^1.5.0"
brotli = "^1.0.9"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
